tests: fix TestDataPlaneWithKonnectExtension
